#3 #43 第3次实验
