+++ students/sec1814080911230/java/RankActivity.java
@@ -2,21 +2,13 @@ package edu.hzuapps.androidlabs.sec1814080911230;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
 
-public class StartGameActivity extends AppCompatActivity {
-
-
+public class RankActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_start_game);
-
-
+        setContentView(R.layout.activity_rank);
     }
 }
