+++ students/sec1814080911230/java/StartGameActivity.java
@@ -1,0 +1,14 @@
+package edu.hzuapps.androidlabs.sec1814080911230;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class StartGameActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_start_game);
+    }
+}
