--- students/sec1814080911230/java/edu/hzuapps/androidlabs/sec1814080911230/RankActivity.java
@@ -1,21 +1,0 @@
-package edu.hzuapps.androidlabs.sec1814080911230;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class RankActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_rank);
-
-        //final Activity thisActivity = this;
-
-    }
-}
